# Removes patch buffer and writes the patch length to the outputStream directly.