--- a/src/com/amazon/ion/impl/bin/IonRawBinaryWriter.java
+++ b/src/com/amazon/ion/impl/bin/IonRawBinaryWriter.java
@@ -694,7 +694,7 @@ import java.util.NoSuchElementException;
 
     private void addPatchPoint(final long position, final int oldLength, final long value)
     {
-        //record the length will be patched in
+        // record the length of the patch
         final int patchLength = WriteBuffer.varUIntLength(value);
         final PatchPoint patch = new PatchPoint(position, oldLength, value);
         if (containers.isEmpty())