*Issue #, if available:*
N/A
*Description of changes:*

This PR introduces an optimization by removing the patch buffer used for storing length data of containers or annotations. The PatchLength and PatchPoint are replaced with the length data. This eliminates the need for writing length data to the patch buffer and simplified the data transfer process. By pass the length information as a field with the PatchPoint, the unnecessary memory allocations will be eliminated, and this change also helps the maintainability of the codebase.

Here is the comparison results from benchmarking a write of data equivalent to the a stream of stream of 194617 nested binary data using IonWriter(binary). The output data will write into an in-memory buffer. (5 forks, 10 iterations, 10 warmups)

**Preallocation 0:**

_Before:_

Benchmark | Score | Error | Units
-- | -- | -- | --
Bench.run | 4455.681 | ± 46.946 | ms/op
Bench.run:Heap usage | 3057.83 | ± 141.196 | MB
Bench.run:Serialized size | 201.663 |   | MB
Bench.run:·gc.alloc.rate | 192.774 | ± 2.753 | MB/sec
Bench.run:·gc.alloc.rate.norm | 935822357 | ± 9234722.854 | B/op
Bench.run:·gc.churn.G1_Eden_Space | 81.197 | ± 2.841 | MB/sec
Bench.run:·gc.churn.G1_Eden_Space.norm | 394264576 | ± 13591199.851 | B/op
Bench.run:·gc.churn.G1_Old_Gen | 44.673 | ± 6.120 | MB/sec
Bench.run:·gc.churn.G1_Old_Gen.norm | 216132485 | ± 28801047.915 | B/op
Bench.run:·gc.churn.G1_Survivor_Space | 1.41 | ± 1.228 | MB/sec
Bench.run:·gc.churn.G1_Survivor_Space.norm | 6912296.75 | ± 6001319.388 | B/op
Bench.run:·gc.count | 132 |   | counts
Bench.run:·gc.time | 47874 |   | ms

_After:_

Benchmark | Score | Error | Units
-- | -- | -- | --
Bench.run | 4353.683 | ± 54.590 | ms/op
Bench.run:Heap usage | 3079.419 | ± 183.491 | MB
Bench.run:Serialized size | 201.663 |   | MB
Bench.run:·gc.alloc.rate | 192.337 | ± 2.324 | MB/sec
Bench.run:·gc.alloc.rate.norm | 912969966 | ± 7548262.441 | B/op
Bench.run:·gc.churn.G1_Eden_Space | 75.176 | ± 4.809 | MB/sec
Bench.run:·gc.churn.G1_Eden_Space.norm | 356515840 | ± 21789150.616 | B/op
Bench.run:·gc.churn.G1_Old_Gen | 45.859 | ± 9.148 | MB/sec
Bench.run:·gc.churn.G1_Old_Gen.norm | 215908789 | ± 41615974.370 | B/op
Bench.run:·gc.churn.G1_Survivor_Space | 1.59 | ± 1.329 | MB/sec
Bench.run:·gc.churn.G1_Survivor_Space.norm | 7668585.81 | ± 6380234.445 | B/op
Bench.run:·gc.count | 117 |   | counts
Bench.run:·gc.time | 44103 |   | ms

**Preallocation 1:**

_Before:_

Benchmark | Score | Error | Units
-- | -- | -- | --
Bench.run | 3991.541 | ± 32.892 | ms/op
Bench.run:Heap usage | 3002.056 | ± 213.998 | MB
Bench.run:Serialized size | 201.663 |   | MB
Bench.run:·gc.alloc.rate | 163.059 | ± 2.099 | MB/sec
Bench.run:·gc.alloc.rate.norm | 713031570 | ± 9233012.903 | B/op
Bench.run:·gc.churn.G1_Eden_Space | 34.671 | ± 3.112 | MB/sec
Bench.run:·gc.churn.G1_Eden_Space.norm | 151596128 | ± 13636978.802 | B/op
Bench.run:·gc.churn.G1_Old_Gen | 54.275 | ± 13.137 | MB/sec
Bench.run:·gc.churn.G1_Old_Gen.norm | 236207049 | ± 55791489.905 | B/op
Bench.run:·gc.churn.G1_Survivor_Space | 2 | ± 1.557 | MB/sec
Bench.run:·gc.churn.G1_Survivor_Space.norm | 8884933.97 | ± 6974512.006 | B/op
Bench.run:·gc.count | 122 |   | counts
Bench.run:·gc.time | 5626 |   | ms

_After:_

Benchmark | Score | Error | Units
-- | -- | -- | --
Bench.run | 3986.978 | ± 49.577 | ms/op
Bench.run:Heap usage | 3133.488 | ± 205.444 | MB
Bench.run:Serialized size | 201.663 |   | MB
Bench.run:·gc.alloc.rate | 161.223 | ± 1.708 | MB/sec
Bench.run:·gc.alloc.rate.norm | 702950841 | ± 9234603.357 | B/op
Bench.run:·gc.churn.G1_Eden_Space | 29.084 | ± 3.111 | MB/sec
Bench.run:·gc.churn.G1_Eden_Space.norm | 126723905 | ± 13264318.356 | B/op
Bench.run:·gc.churn.G1_Old_Gen | 49.225 | ± 12.427 | MB/sec
Bench.run:·gc.churn.G1_Old_Gen.norm | 213226636 | ± 52205386.688 | B/op
Bench.run:·gc.churn.G1_Survivor_Space | 1.789 | ± 1.172 | MB/sec
Bench.run:·gc.churn.G1_Survivor_Space.norm | 7797404.43 | ± 5079736.242 | B/op
Bench.run:·gc.count | 105 |   | counts
Bench.run:·gc.time | 4323 |   | ms


**Preallocation 2:**

Before:

Benchmark | Score | Error | Units
-- | -- | -- | --
Bench.run | 3991.828 | ± 25.325 | ms/op
Bench.run:Heap usage | 2587.51 | ± 105.583 | MB
Bench.run:Serialized size | 204.76 |   | MB
Bench.run:·gc.alloc.rate | 161.335 | ± 2.048 | MB/sec
Bench.run:·gc.alloc.rate.norm | 705458559 | ± 7538111.558 | B/op
Bench.run:·gc.churn.G1_Eden_Space | 38.321 | ± 2.435 | MB/sec
Bench.run:·gc.churn.G1_Eden_Space.norm | 167660312 | ± 10798782.611 | B/op
Bench.run:·gc.churn.G1_Old_Gen | 82.543 | ± 6.432 | MB/sec
Bench.run:·gc.churn.G1_Old_Gen.norm | 361075125 | ± 27841404.073 | B/op
Bench.run:·gc.churn.G1_Survivor_Space | 1.282 | ± 2.294 | MB/sec
Bench.run:·gc.churn.G1_Survivor_Space.norm | 5631533.01 | ± 10088275.751 | B/op
Bench.run:·gc.count | 144 |   | counts
Bench.run:·gc.time | 1562 |   | ms

After:

Benchmark | Score | Error | Units
-- | -- | -- | --
Bench.run | 3932.218 | ± 35.715 | ms/op
Bench.run:Heap usage | 2641.364 | ± 99.636 | MB
Bench.run:Serialized size | 204.76 |   | MB
Bench.run:·gc.alloc.rate | 163.452 | ± 2.999 | MB/sec
Bench.run:·gc.alloc.rate.norm | 705397751 | ± 7537218.072 | B/op
Bench.run:·gc.churn.G1_Eden_Space | 38.229 | ± 2.882 | MB/sec
Bench.run:·gc.churn.G1_Eden_Space.norm | 164947995 | ± 12079327.544 | B/op
Bench.run:·gc.churn.G1_Old_Gen | 79.123 | ± 7.516 | MB/sec
Bench.run:·gc.churn.G1_Old_Gen.norm | 341289885 | ± 31607423.749 | B/op
Bench.run:·gc.churn.G1_Survivor_Space | 1.836 | ± 1.872 | MB/sec
Bench.run:·gc.churn.G1_Survivor_Space.norm | 7931019.31 | ± 8002315.605 | B/op
Bench.run:·gc.count | 146 |   | counts
Bench.run:·gc.time | 1502 |   | ms

For more benchmark results generated from other dataset, please visit [here](https://gist.github.com/linlin-s/84aed46a96c4fd6f0ce61669aab95d7f).




By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.