@linlin-s 2023-07-21
Here are some new benchmark results from different benchmark usage pattern:

Results from benchmarking a write of data equivalent to a stream of 59155 nested binary Ion values(3 forks, 2 warmups, 2 iterations).


Benchmark | No Change | Remove Patch Buffer | Units
-- | -- | -- | --
Bench.run | 509.652 | 503.944 | ms/op
Bench.run:Heap usage | 326.793 | 355.376 | MB
Bench.run:Serialized size | 21.271 | 21.271 | MB
Bench.run:·gc.alloc.rate | 125.438 | 126.443 | MB/sec
Bench.run:·gc.alloc.rate.norm | 70369781.67 | 70038249.64 | B/op
Bench.run:·gc.churn.G1_Eden_Space | 7.477 | 5.494 | MB/sec
Bench.run:·gc.churn.G1_Eden_Space.norm | 4194304 | 3043367.01 | B/op
Bench.run:·gc.churn.G1_Old_Gen | 133.457 | 132.904 | MB/sec
Bench.run:·gc.churn.G1_Old_Gen.norm | 74868326.4 | 73639994.51 | B/op
Bench.run:·gc.churn.G1_Survivor_Space | 0.03 | 0.029 | MB/sec
Bench.run:·gc.churn.G1_Survivor_Space.norm | 16604.8 | 16085.956 | B/op
Bench.run:·gc.count | 78 | 73 | counts
Bench.run:·gc.time | 585 | 518 | ms

Results from benchmarking a write of data equivalent to a single nested binary Ion value(3 forks, 2 warmups, 2 iterations).


Benchmark | No Change | Remove Patch Buffer | Units
-- | -- | -- | --
Bench.run | 0.089 | 0.082 | ms/op
Bench.run:Heap usage | 6.778 | 8.811 | MB
Bench.run:Serialized size | 0.005 | 0.005 | MB
Bench.run:·gc.alloc.rate | 93.704 | 96.721 | MB/sec
Bench.run:·gc.alloc.rate.norm | 9172.312 | 8701.888 | B/op
Bench.run:·gc.churn.G1_Eden_Space | 94.339 | 97.196 | MB/sec
Bench.run:·gc.churn.G1_Eden_Space.norm | 9234.427 | 8745.584 | B/op
Bench.run:·gc.churn.G1_Survivor_Space | 0.004 | 0.004 | MB/sec
Bench.run:·gc.churn.G1_Survivor_Space.norm | 0.387 | 0.317 | B/op
Bench.run:·gc.count | 574 | 592 | counts
Bench.run:·gc.time | 317 | 309 | ms

Results from benchmarking a write of data equivalent to a stream of 194617 nested binary Ion value(3 forks, 2 warmups, 2 iterations).


Benchmark | No change | Remove Patch Buffer | Units
-- | -- | -- | --
Bench.run | 3988.343 | 3974.707 | ms/op
Bench.run:Heap usage | 2815.305 | 2654.546 | MB
Bench.run:Serialized size | 201.663 | 201.663 | MB
Bench.run:·gc.alloc.rate | 160.884 | 160.583 | MB/sec
Bench.run:·gc.alloc.rate.norm | 703010831.1 | 713050404 | B/op
Bench.run:·gc.churn.G1_Eden_Space | 32.785 | 40.088 | MB/sec
Bench.run:·gc.churn.G1_Eden_Space.norm | 143072369.8 | 178374428.4 | B/op
Bench.run:·gc.churn.G1_Old_Gen | 58.898 | 73.473 | MB/sec
Bench.run:·gc.churn.G1_Old_Gen.norm | 255793294.2 | 326050275.6 | B/op
Bench.run:·gc.churn.G1_Survivor_Space | 6.423 | 4.629 | MB/sec
Bench.run:·gc.churn.G1_Survivor_Space.norm | 28208808 | 20096732 | B/op
Bench.run:·gc.count | 16 | 20 | counts
Bench.run:·gc.time | 571 | 788 | ms